+++ apps/deepl_integration/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DeeplIntegrationConfig(AppConfig):
+    name = 'deepl_integration'
